+++ lib/subscription_impl.rb
@@ -4,7 +4,6 @@
 require_relative "subscription_impl/version"
 require 'node'
 require 'node_diagram'
-require 'node_subscription_impl'
 
 # Subscription.
 # @class_description